requirements.txt
altgraph==0.17.4
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
colorama==0.4.5
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.9
Django==4.0.6
django-admin-interface==0.23.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-recaptcha==3.0.0
et-xmlfile==1.1.0
filelock==3.17.0
fonttools==4.36.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==1.1.1
MarkupPy==1.14
mysqlclient==2.1.1
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==24.0
pandas==1.5.1
pefile==2023.2.7
Pillow==9.2.0
platformdirs==4.3.6
pluggy==1.0.0
psycopg==3.2.4
py==1.11.0
pycparser==2.21
pydyf==0.2.0
pyinstaller-hooks-contrib==2024.3
PyMySQL==1.0.2
pyparsing==3.0.9
pyphen==0.12.0
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.5
pywin32-ctypes==0.2.2
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tinycss2==1.1.1
tomli==2.0.1
tornado==6.4
typing_extensions==4.12.2
tzdata==2022.1
virtualenv==20.29.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
